# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AgentMode = (typeof AgentMode)[keyof typeof AgentMode];

export const AgentMode = {
  react: "react",
  pipeline: "pipeline",
} as const;

export interface Agent {
  id: number;
  slug: string;
  name: string;
  description: string;
  systemPrompt: string;
  tools: string[];
  mode: AgentMode;
  model: string;
  createdAt: string;
  updatedAt: string;
  runCount: number;
}

export type CreateAgentBodyMode =
  (typeof CreateAgentBodyMode)[keyof typeof CreateAgentBodyMode];

export const CreateAgentBodyMode = {
  react: "react",
  pipeline: "pipeline",
} as const;

export interface CreateAgentBody {
  name: string;
  description: string;
  systemPrompt: string;
  tools: string[];
  mode?: CreateAgentBodyMode;
  model?: string;
}

export type UpdateAgentBodyMode =
  (typeof UpdateAgentBodyMode)[keyof typeof UpdateAgentBodyMode];

export const UpdateAgentBodyMode = {
  react: "react",
  pipeline: "pipeline",
} as const;

export interface UpdateAgentBody {
  name?: string;
  description?: string;
  systemPrompt?: string;
  tools?: string[];
  mode?: UpdateAgentBodyMode;
  model?: string;
}

export type AgentRunStatus =
  (typeof AgentRunStatus)[keyof typeof AgentRunStatus];

export const AgentRunStatus = {
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface AgentRun {
  id: number;
  agentId: number;
  agentName: string;
  task: string;
  status: AgentRunStatus;
  output?: string;
  stepCount: number;
  createdAt: string;
  completedAt?: string;
  /** When this run was created via replay, the id of the source run. */
  replayOfRunId?: number | null;
  /** When set, this run is publicly viewable as a Strand at /strand/{token}. */
  shareToken?: string | null;
}

export type RunStepType = (typeof RunStepType)[keyof typeof RunStepType];

export const RunStepType = {
  thought: "thought",
  tool_call: "tool_call",
  tool_result: "tool_result",
  final_answer: "final_answer",
} as const;

export interface RunStep {
  id: number;
  runId: number;
  stepIndex: number;
  type: RunStepType;
  content: string;
  toolName?: string;
  createdAt: string;
}

export type AgentRunWithStepsStatus =
  (typeof AgentRunWithStepsStatus)[keyof typeof AgentRunWithStepsStatus];

export const AgentRunWithStepsStatus = {
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface AgentRunWithSteps {
  id: number;
  agentId: number;
  agentName: string;
  task: string;
  status: AgentRunWithStepsStatus;
  output?: string;
  stepCount: number;
  createdAt: string;
  completedAt?: string;
  /** When this run was created via replay, the id of the source run. */
  replayOfRunId?: number | null;
  /** When set, this run is publicly viewable as a Strand at /strand/{token}. */
  shareToken?: string | null;
  steps: RunStep[];
}

export type ReplayRunResponseStatus =
  (typeof ReplayRunResponseStatus)[keyof typeof ReplayRunResponseStatus];

export const ReplayRunResponseStatus = {
  completed: "completed",
  failed: "failed",
} as const;

export interface ReplayRunResponse {
  runId: number;
  agentId: number;
  replayOfRunId: number;
  status: ReplayRunResponseStatus;
  output: string;
}

export interface StrandShareResponse {
  token: string;
  url: string;
}

export type StrandAgentMode =
  (typeof StrandAgentMode)[keyof typeof StrandAgentMode];

export const StrandAgentMode = {
  react: "react",
  pipeline: "pipeline",
} as const;

export type StrandStatus = (typeof StrandStatus)[keyof typeof StrandStatus];

export const StrandStatus = {
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

/**
 * Public, read-only view of a single agent run trace.
 */
export interface Strand {
  id: number;
  token: string;
  agentId: number;
  agentName: string;
  agentMode: StrandAgentMode;
  task: string;
  status: StrandStatus;
  output?: string;
  stepCount: number;
  createdAt: string;
  completedAt?: string;
  steps: RunStep[];
}

export type AgentTemplateDifficulty =
  (typeof AgentTemplateDifficulty)[keyof typeof AgentTemplateDifficulty];

export const AgentTemplateDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface AgentTemplate {
  id: string;
  name: string;
  description: string;
  category: string;
  systemPrompt: string;
  tools: string[];
  model: string;
  difficulty: AgentTemplateDifficulty;
}

export interface PlatformStats {
  totalAgents: number;
  totalRuns: number;
  completedRuns: number;
  failedRuns: number;
  activeTools: string[];
}

export interface StartRunBody {
  task: string;
}

export type ToolType = (typeof ToolType)[keyof typeof ToolType];

export const ToolType = {
  webhook: "webhook",
  prompt: "prompt",
} as const;

export type ToolConfig = { [key: string]: unknown };

export interface Tool {
  id: number;
  name: string;
  description: string;
  type: ToolType;
  config: ToolConfig;
  createdAt: string;
  updatedAt: string;
}

export type CreateToolBodyType =
  (typeof CreateToolBodyType)[keyof typeof CreateToolBodyType];

export const CreateToolBodyType = {
  webhook: "webhook",
  prompt: "prompt",
} as const;

export type CreateToolBodyConfig = { [key: string]: unknown };

export interface CreateToolBody {
  name: string;
  description: string;
  type: CreateToolBodyType;
  config: CreateToolBodyConfig;
}

export interface BuiltinTool {
  id: string;
  name: string;
  description: string;
  category: string;
}

export interface DeleteToolParams {
  id: number;
}