# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentBodyMode } from "./createAgentBodyMode";

export interface CreateAgentBody {
  name: string;
  description: string;
  systemPrompt: string;
  tools: string[];
  mode?: CreateAgentBodyMode;
  model?: string;
}